1)Finished PlayerPanel and StationsList
