+++ my-blog-react/src/App.scss
@@ -9,7 +9,10 @@
   height: 100%;
 }
 
-
+.clearPaper {
+  background: inherit;
+  box-shadow: none;
+}
