+++ my-blog-react/src/Redux/Reducers/MenuSection-reducer.js
@@ -3,10 +3,10 @@ const SOME_ACTION_TYPE = 'SOME_ACTION_TYPE';
 const initialState = {
      sections: [
           { title: 'Main', url: '/main' },
+          { title: 'Music', url: '/music' },
           { title: 'Posts', url: '/posts' },
           { title: 'Usefull links', url: '/usefull' },
           { title: 'Feed back', url: '/feedback' },
-          { title: 'Just take a rest', url: '/rest' },
           { title: 'Contacts', url: '/contacts' },
      ],
